Updated NationalRail to Python 3

+++ NationalRail.py
@@ -2,10 +2,9 @@
 # Project Website : https://departureboard.jonathanfoot.com
 # Documentation   : https://jonathanfoot.com/Projects/DepartureBoard
 # Description     : This program allows you to display a live train departure board for any UK station nationally, excluding the London Underground.
+# Python 2 Required (Deprecated)
 
-import urllib2
 import time
-import math
 import inspect,os
 import sys
 import inflect
@@ -13,10 +12,8 @@ import re
 import argparse
 from PIL import ImageFont, Image, ImageDraw
 from luma.core.render import canvas
-from luma.core.interface.serial import spi
-from luma.core import cmdline, error
-from lxml import objectify
-from datetime import datetime, date
+from luma.core import cmdline
+from datetime import datetime
 from luma.core.image_composition import ImageComposition, ComposableImage
 from nredarwin.webservice import DarwinLdbSession
 
@@ -68,6 +65,7 @@ parser.add_argument("--FixNextToArrive",dest='FixToArrive', action='store_true',
 parser.add_argument('--no-splashscreen', dest='SplashScreen', action='store_false',help="Do you wish to see the splash screen at start up; recommended and on by default.")
 parser.add_argument("--Display", default="ssd1322", choices=['ssd1322','pygame','capture','gifanim'], help="Used for development purposes, allows you to switch from a physical display to a virtual emulated one; default 'ssd1322'")
 parser.add_argument("--max-frames", default=60,dest='maxframes', type=check_positive, help="Used only when using gifanim emulator, sets how long the gif should be.")
+parser.add_argument("--Acknowledge-Deprecated", dest='acknowledgeDep', action='store_true', help="ATTENTION - You are acknowledging you understand this version of the program is deprecated and is not supported anymore. Doing this will hide the update message, but does mean you won't get any more updates.")
 
 # Defines the required paramaters
 requiredNamed = parser.add_argument_group('required named arguments')
@@ -735,7 +733,10 @@ StartUpDate = datetime.now().date()
 # Draws the clock and tells the rest of the display next frame wanted.
 def display():
     board.tick()
-    msgTime = str(datetime.now().strftime("%H:%M:%S" if (Args.TimeFormat==24) else "%I:%M:%S"))	
+    if Args.acknowledgeDep:
+	    msgTime = str(datetime.now().strftime("%H:%M:%S" if (Args.TimeFormat==24) else "%I:%M:%S"))	
+    else:
+	    msgTime = 'update.jonathanfoot.com' 
     with canvas(device, background=image_composition()) as draw:
         image_composition.refresh()
         draw.multiline_text((0, 0), board.GetHeader(), font=BasicFont)
@@ -751,6 +752,7 @@ def Splash():
 
 
 try:
+    print "\033[1;31m DEPRECATED WARNING: THIS VERSION OF THE SOFTWARE IS NO LONGER SUPPORTED, IT IS STRONGLY RECOMMEND YOU MANUALLY UPGRADE IT. MORE INFORMATION CAN BE FOUND AT UPDATE.JONATHANFOOT.COM\033[0;0m"
     Splash() 
  	# Run the program forever		       
     while True:
@@ -764,7 +766,7 @@ try:
 		# Turns the display into one of the energy saving modes if in the correct time and enabled.
         if (Args.EnergySaverMode != "none" and is_time_between()):
             # Check for program updates and restart the pi every 'UpdateDays' Days.
-            if (datetime.now().date() - StartUpDate).days >= Args.UpdateDays:
+            if (datetime.now().date() - StartUpDate).days >= Args.UpdateDays and not Args.acknowledgeDep:
                 print "Checking for updates and then restarting Pi."
                 os.system("sudo git -C %s pull; sudo reboot" % (os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))))
                 sys.exit()
